The "max ticks if won" was a red herring, fixed.
The real problem was that the PFB-cat had its costs all wrong and never
behaved properly.

I have now gone by the assumption that cost function of the original
thesis had a very different scale than mine, so the hill cost was
scaled correctly. I set it to a hundreth (10), as this seems to give
good behaviour versus the follower dog ai.

In doing so I also removed the "scaling factor" I had applied to  the
exit cost which is now irrelevant. Hopefully this will finally lay to
rest the problems we are having in the simulation.